+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Revamped to appd-exts-commons 2.1.0
 * Moved to Asynchronous processing using FutureTasks
 
-### Version 2.2.0-SNAPSHOT
+### Version 4.0.0
 
-* Revamped to appd-exts-commons 2.2.0-SNAPSHOT
+* Revamped to appd-exts-commons 2.2.0
+* supports for composite lang memory metrics for Hbase